requirements.txt
ujson
beard
munkres
editdistance
repoze.lru

